requirements.txt
easygui==0.98.3
python-dotenv==1.0.0